Stand Ritrovo up on the released kernel in one command
Ritrovo installed correctly before this, but only for someone willing to walk
nine manual steps in docs/INSTALL.md against a Trovato they built themselves.
Nothing proved it still stood up on a published release, and the last time
anyone had checked, the kernel was about thirty commits older than 0.101.0.

    scripts/serve-demo.sh

Docker is now the only prerequisite. docker-compose.demo.yml brings up Postgres
and Redis, compiles the five plugins to WebAssembly in a throwaway
rust:1-bookworm container, starts ghcr.io/jeremyandrews/trovato:0.101.0 with the
Ritrovo overlay appended to its three search paths, and pokes /cron until the
conference import queue drains. `docker compose -f docker-compose.demo.yml up`
reaches the same site; serve-demo.sh is the wrapper that waits for it and checks
the result.

The ordering is the whole difficulty, and it lives in scripts/demo-bootstrap.sh:
serve briefly (kernel migrations, plugin discovery), complete the installer over
HTTP, stop, import the config set, enable the plugins, import the content, then
exec the real server. Config before enable is the trap: ritrovo_importer resolves
the topic taxonomy once, in tap_install, which fires at the first start after the
plugin is enabled. Get it backwards and every conference imports untagged. On a
correct run the log reads `discover_taxonomy_uuids: 23/23 terms found`.

Verified against 0.101.0 on 2026-08-21, empty Docker to checked site in 1m41s:
five plugins enabled, 269 import batches drained, 5,644 conferences (160
upcoming), both importer admin screens answering 401 anonymous / 200
administrator, /search?q=rust returning 37 results, /it/conferenze rendering with
lang="it". No api_version change was needed: 0.99 plugins are accepted by a 0.101
kernel, because the rule is plugin minor <= kernel minor.

Two decisions worth reading the reasoning for rather than inferring:

  * docs/tutorial/templates is now vendored, nine files copied from the release.
    The image ships templates/, static/ and docs/tutorial/config/ and stops
    there, so a demo that runs on the image alone has to get the tutorial's
    templates from somewhere. Mounting them needs the checkout the demo exists to
    do without, and fetching them at start-up puts a network failure mode between
    a stranger and their first look. The cost is drift, and it is paid:
    scripts/check-tutorial-templates.sh diffs them against the release in CI.
    docs/INSTALL.md records the tradeoff, reversing what that file used to argue.

  * demo/config carries three url_alias entries the tutorial set already has,
    stored without their /it/ prefix. The kernel's language middleware strips the
    prefix before the alias lookup runs, so the prefixed rows can never match and
    /it/conferenze 404s on a stock install while /it/conferences answers 200.
    That mismatch is Trovato's, between its own config set and its own
    middleware; these rows are the overlay-side way to have the Italian URLs the
    tutorial advertises actually resolve. demo/config/README.md has the detail.

Also here: CI, which the README has wanted for a while. It gates the wasm build
against the pinned SDK revision, fmt and clippy with warnings as errors,
shellcheck, and demo/checks — a test crate that pins the demo's wiring to the
repository it demonstrates, so a sixth plugin nobody adds to the enable list
fails a test instead of quietly not appearing. One pre-existing clippy warning in
ritrovo_importer is fixed so that gate can be warnings-as-errors.

+++ plugins/ritrovo_importer/src/lib.rs
@@ -241,7 +241,7 @@ pub fn tap_install() -> serde_json::Value {
 
     // 2. Queue historical import.
     let now = current_timestamp();
-    let current_year = timestamp_to_year(now) as u16;
+    let current_year = timestamp_to_year(now);
     let mut pushed = 0u32;
     let mut errors = 0u32;
